project_euler/pyproject.toml
[project]
name = "project-euler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mathlib",
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
]

[tool.uv.sources]
mathlib = { workspace = true }
